Implement working policy enforcement and provisioning.
In the process, rework some state storage methods. For the file storage backend,
store the state files in a user directory to allow a (as yet non-existant) UI for
allowing the user and/or admin to request a remote device wipe. Also seperate out
the device-specific information such as the policykey, user-agent, device string etc...
into it's own state file. Again, this will be more efficient in a history/sql based state
storage driver.