More improvements to provisioning, fix variable names.
With this and the last commit, provisioning can now be set to Allow,
while still allowing Android devices to connect. The Android devices will
not be provisioned, so setting it to 'Force' will still prevent them from
connecting, but now at least using Android won't prevent allowing other
devices to be provisioned.