Ignore duplicate incoming message additions (Bug: 10644)
If incoming changes are successfully applied, but the client does
not receive (or ignores) the new synckey, the next SYNC command will
send duplicate client changes. We now save the incoming clientid
string in the map table so we can ignore duplicate additions in these cases.

Note the new migration that must be run