Fix generating ActiveSync Timezone structures.
When Horde_Date receives an iso8601 style datetime string, it uses
date_create(), and force changes the timezone to date_default_timezone_get().
Specifying 'UTC' as the 2nd parameter to Horde_Date doesn't work, the tz parameter
is always ignored in this case. We must change the timezone explicitly back to UTC
after creating the Horde_Date object in this way. Of course, there is also
another bug that prevents this from working properly on the DST -> Standard switch
(see broken test), so we have to use DateTime directly in that case.

I've tried numerous ways of allowing the tz parameter to Horde_Date
be honored in this case, but there was always some case that wouldn't
work properly. Fix one, break another.