Renaming IMAP folders via EAS not supported.
In Exchange, each mailbox has a GUID that the device tracks. Renaming
the exchange mailbox changes the display name, but not the GUID. EAS
clients do not support changing the UID. Since IMAP mailboxes have no
such identifier, we cannot support changing the folder on the device.

Changing on the server will send a REMOVE and ADD command that has the
net effect of removing the old mailbox name and adding the new name. Even
if we force the server to send a REMOVE/ADD after changing on the client
the client still holds on to the old GUID - which leads to polling the
IMAP server for the incorrect maibox.