Persist the last known synckey.
Keep the last known synckey for each collection regardless of the
SYNCKEY value from the request. When manually triggering a SYNC request
(like from clicking a refresh button) some devices will send a full SYNC
request with only the one collection instead of a PARTIAL SYNC. This
breaks using the syncCache for PING requests, as a PING request MUST have
a previous synckey.