Apple does it again. Handle protocol violation that causes iOS devices to crash.
Apple ignores the specs and sends a COMMANDS container with a synckey of 0. Granted,
it's only asking for a FETCH which should theoretically be possible without valid state,
but then the device crashes if we actually try to honor the FETCH request.

Ignore all COMMANDS containers in an initial SYNC request (synckey == 0).