Validating timestamps simply cannot work for PING requests.
Since some devices will both send mulitiple PING requests as well
as SYNC requests while a PING is still alive, the timestamp validation
can never be counted on to be accurate. This is not an issue for clients
that truly implement the 12.1 hanging SYNC since it all takes place in
the same SYNC request.