Bug: 11635 Yet Another fix for broken ActiveSync clients.
Some android clients not only incorrectly send the email address
as the username in an Authorization header, but they also completely
fail to send the XML Autodiscover request. Try to recover from this
by getting the email from the Authorization header and hope for the best.

Note, these same clients will incorrectly auto populate the username
field with the email address.