Better way of dealing with broken email charsets.
Take a cue from Horde_Mime_Headers:: and try the more common mulitbyte
charsets before giving up. Try removing non 7-bit characters if we still
can't get good UTF-8 and, if for whatever reason, THAT fails, force
a conversion from the supposed charset to UTF-8 and hope for the best.