Have Horde_ActiveSync_Request_SendMail sniff out the type of request.
In EAS 14.x the Sendmail command may or maynot be transmitted as
either IMF/Rfc822 stream or as a wbxml encoded stream of COMPOSEMAIL
namespace properties. Check for valid wbxml and if none is present,
assume the full input stream is rfc822 input.

Still need to implement the wbxml decoding for the SendMail request.