Some clients send the SYNC_FOLDERTYPE data in EAS > 12.1 here.
When sending SYNC_ADD commands, *some* clients will send the SYNC_FOLDERTYPE
node here when EAS > 12.1 while others will not. So far, Outlook is the
only client I have seen that does this, so we need to be able to accept it
if it's sent.

Fixes being unable to add new items from Outlook.