Fix sending deletions to the client an item was created on in certain cases.
If an item created on one EAS client is deleted from another EAS client
before the first client goes through another full SYNC cycle,
the deletion was being ignored since an entry still existed in
the map table for that UID/CLIENT_ID combination.

Fix this by tracking if a map entry is a deletion, and save some
history calls to boot.

Note that a new migration is added with this commit.