Don't filter out FLAG_DELETED messages.
This can cause all kinds of problems in supporting this for both
CONDSTORE and non-CONDSTORE servers. If using ActiveSync and a MUA
that does NOT use a Trash folder (as EAS expects), then the trade
off is that messages flagged as \DELETED will still appear as
"normal" on the EAS client.