Another place to work around broken clients.
Broken clients like e.g., iOS, send a request to PING a collection
before the collection is synchronized. This is a violation of the
protocol (see MS-ASCMD 3.1.5.6). To work around this, continue
to mark the collection as "pingable", but ignore the collection
(and the exception that is thrown when we try to initialize the
state).

Possibly "fixes" Bug: 12605.