Another fix for Bug: 12934.
Wonderful. Outlook sends the collection id as the *calendar* id, not
the email folder's id when sending meeting responses regardless of
which interface the response is sent from, email or calendar. Thank
goodness that it at least keeps the original message UID so we can
still find it in the INBOX.

So, we now have to always retry with the INBOX collection if fetching
the response fails. If the user moves the invitation out of the INBOX
before responding to it, we don't really have any chance of finding it
so must give up.