Do not prevent a DELETE command from being sent to the client after a MOVEITEMS.
It seems that after a MOVEITEMS command, we aren't supposed to mute the mirroring back
of the removal of the src message. Some clients rely on this to remove the original message
from the UI. This fixes the annoying issues in OL2014 that were causing duplicate messages
to appear whenever an item was moved/deleted.