Another place to catch SMS changes.
If SMS sync was enabled at some point on the client, and then disabled, the
client is *supposed* to remove any already existing SMS messages from
the client's inbox. If it doesn't, and the user deletes them,
we need to catch it before it is attempted to be sync'd to the IMAP server.