Bug: 14086  Work around various broken PHP behaviours regarding stream filters.
It appears to be safe to use the base64-encode filter when it's attached
as a STREAM_FILTER_WRITE. The behavior is only broken when the stream is
being read through the filter, so perform the filtering in this method
instead of just attaching the filter to the stream that is passed to the
wbxml encoder. Adds a small memory overhead but this now works reliably
with packages such as Debian's that have removed the BC breaking PHP
patch as well as with PHP binaries that have the change included.