The correct place to work around this bug is in Horde_Mime.
We need to perform this without performing a transfer decode.
Since using Horde_Mime_Part::getContents() always returns
the decoded/binary content, we can't use that value to call
setContents() and expect the transfer encoding to be the same.

In some cases, when the original content was base64, and the version
of PHP in use causes the base64 decode stream filter to error out
on non-base64 characters, this would mangle text parts.

Also, why is this PHP bug still open after over 2 years, and an
available patch?