Add a DATE_TYPE_LOCAL.
As usual, MS docs are not clear on how to set these values. They state
that they are DateTime values (and must be sent as UTC), but then
they have a "local" vs "UTC" property for these task dates. To
complicate things, different clients interpret the date values
differently.

We are going to go with what Outlook is sending/expecting to simplify
things. That is, the utcstartdate/utcduedate is the "normal" date field that
is the local date converted to UTC. The startdate (which is described
in the docs as the "local date"). Is the local time, but it is sent
as a datetime string that looks like a UTC time (with the Z) at the end.

*sigh*