Handle this as a temporary server error.
After thinking about this, it's better to handle this as a server error
and let the client try again later a few times before failing.
Otherwise, this could lead to constant resetting/resyncing of the
client and causing high load on the server. Also, this way, if
the client is well behaved, it will notify the user of the failure
instead of just failing to add new email.