Catch Horde_Idna_Exception.
This exception is leaked up from Horde_Mail through Horde_Mime and
it looks like this was intentional. Either way, there is code that at
least in IMP that currently relies on this, so we can't change this
since it would break BC.

This fixes letting an email with an illegal domain label from breaking
sync all together.