Fix broken handling of PING requests when folder state is missing.
This was broken in that if it was the result of a "full" PING request,
it would force a SYNC, but would not indicate which folders needed
to by SYNCd. This reverts back to the previous behaviour where any
missing state discovered during a PING request will result in that
folder being flagged for a SYNC normally, where the SYNC request
will catch the missing state and force a SYNC_KEY of 0.

Discovered when IMP maint. task renamed my Sent Mail folder... :)