Try to honor BODYPREF when we don't have an HTML part.
If we only have a plain part, but BODYPREF requests HTML,
convert to HTML so we can honor this. Some clients don't honor
the TYPE property being set to PLAIN if HTML was requested,
even though this is to be taken as only a preference, not a
MUST.

Fixes broken behavior in GMail app on Android that would cause
reply text to be lost when replying to a plain text email.