Fix issue with handling TNEF attachments;
The used to work on older EAS versions because it would
return the entire base mime part. With modern versions
we only return a list of attachments initially, then fetch
the attachement data only if needed, identified by the mime part
id. This means we need to rewrite the basePart's structure any
time we _might_ be tryting to access the mime part of a tnef
attachment.